Refactor playlist saving functionality and improve queue management
- Simplified the process of saving the current queue as a private playlist by moving the logic from SongActionsHelper to PlayerService.
- Updated QueueFooterActions to streamline the save playlist action, removing unnecessary parameters.
- Enhanced the queue duplicate check by using a more concise syntax.
- Removed the outdated saveQueueAsPlaylist method from SongActionsHelper to improve code clarity and maintainability.
